+++ CHANGELOG.md
@@ -5,6 +5,10 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+- audit logs: export support
+- client: streaming Get/Do variants
+
 ## [8.19.0]
 
 ### Added